fix: there was changed approach for pages routing (People, Colors)
