+++ src/screens/Colors/ColorsList/ColorsList.module.scss
