+++ src/screens/People/People.jsx
@@ -1,19 +1,12 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import styles from './People.module.scss';
-import PersonCard from "./PersonCard/PersonCard";
 import { Outlet, useNavigate, useParams } from "react-router-dom";
 import Pagination from '../../components/shared/Pagination/Pagination';
 import { useSelector } from "react-redux";
-import { ROUTES } from "../../enums/ROUTES";
-import { sessionController } from "../../services/storage/sessionController";
 
 const People = () => {
-  const [people, setPeople] = useState([]);
-
   const navigate = useNavigate();
   const routeParams = useParams();
-  const peopleCollection = useSelector(state => state.peopleCollection.people);
-  const peopleFromServer = useSelector(state => state.peoplePage.people);
   const totalPages = useSelector(state => state.peoplePage.pagesAmount);
 
   useEffect(() => {
@@ -22,50 +15,7 @@ const People = () => {
     }
   }, [routeParams])
 
-  useEffect(() => {
-    const filteredArray = filterByDeletedPeople();
-    setPeople(filteredArray);
-    filterByUser();
-  }, [peopleFromServer]);
-
-  function filterByDeletedPeople() {
-    return peopleCollection.filter(personFromDB => {
-      return peopleFromServer.find(personFromServer => personFromServer.id === personFromDB.id);
-    })
-  }
-
-  function filterByUser() {
-    if (Number(totalPages) === Number(routeParams.page)) {
-      const userFromStorage = sessionController.getUser();
-
-      const isUserExistInPeopleCollection = peopleCollection.find(person => person.id === userFromStorage.id);
-      if (!isUserExistInPeopleCollection) {
-        setPeople(people => people.concat([userFromStorage]));
-      }
-    }
-  }
   
-  function increasePageNumber() {
-    const nextPage = Number(routeParams.page) + 1;
-    navigate(`${ROUTES.people}/${nextPage}`, {replace: false});
-  }
-
-  function decreasePageNumber() {
-    const prevPage = Number(routeParams.page) - 1;
-    navigate(`${ROUTES.people}/${prevPage}`, {replace: false});
-  }
-
-  function handleChosenPage(num) {
-    navigate(`${ROUTES.people}/${num}`, {replace: false});
-  }
-
-  function openPersonDetailsPage(id) {
-    navigate(`${ROUTES.person}/${id}`);
-  }
-
-
-
-
   function goToNextPage() {
     const nextPage = Number(routeParams.page) + 1;
     navigate(`${nextPage}`, {replace: false});
@@ -80,21 +30,6 @@ const People = () => {
     navigate(`${num}`, {replace: false});
   }
 
-
-
-
-
-
-
-
-  // const renderUsers = people.map(user => {
-  //   return (
-  //     <div className={styles['users-page__user']} key={user.id.toString()}>
-  //       <PersonCard user={user} onCardClick={openPersonDetailsPage} />
-  //     </div>
-  //   )
-  // })
-
   return (
     <div className={styles['users-page']}>
       <h1 className={styles['users-page__title']}>People</h1>
@@ -102,7 +37,6 @@ const People = () => {
 
       <div className={styles['users-page__users']}>
         <Outlet />
-        {/* {<ListPersonCards />} */}
       </div>
       <div className={styles['users-page__pagination-wrapper']}>
         <Pagination
