+++ src/screens/People/PeopleList/PeopleList.module.scss
